Added patch to allow a final_filter function that is applied to combo URLs
This is a function that takes the final combo URL and applies an
arbitrary function to it before requesting it. This allows cool things
like compressing the full URL and other interesting things. Since full
control is given to the final_filter function declared in the YUI config,
this ignores the normal max URL length splitting.